병합전 정리

+++ Source/Base/Public/FpsCppProjectile.h
@@ -53,6 +53,6 @@ public:
 
 
 private:
-	int32 Damage;
+	float Damage;
 };
